refactor(consumer): split main.go into smaller separate files
this moves out error handler, handlers to interact with db, and any
other relevant logic into separate files.
