+++ consumer/error.go
@@ -1,0 +1,9 @@
+package main
+
+import "log"
+
+func failOnError(err error, msg string) {
+	if err != nil {
+		log.Panicf("%s: %s", msg, err)
+	}
+}
