+++ consumer/internal/model.go
@@ -1,0 +1,7 @@
+package internal
+
+type DeviceData struct {
+	DeviceID    string  `json:"device_id"`
+	Temperature float64 `json:"temperature"`
+	Humidity    float64 `json:"humidity"`
+}
